An improved version of ddemangle.
This one uses std.regex to replace all occurrences of mangled symbols
per line. Sometimes this is needed when the line just happens to contain
the wrong character sequences (_D...), which cause a real symbol later
on in the line not to be demangled, if only the first match is
attempted.

It also uses std.getopt for future extensions (ala c++filt, if we ever
decide to do that).