Do not exclude tango package from compilation
Considering D2 tango uses druntime this is not helpful. Quite the contrary, that force me to use custom rdmd build to test my D1->D2 in-progress ports (which include necessary parts of tango in-line).

I think it is quite reasonable to ask projects using Tango-D2 to specify exclusion manually in their build scripts.