Add rdmdArguments helper function to rdmd_test
The existing rdmd_test script assumes that the compiler invoked by rdmd
when running tests is the same as the compiler used to build rdmd, and
that it targets the same build model.  This helper function should make
it easier to construct tests that use different compiler and/or model
settings.