Add optional --test-compilers flag to rdmd_test
This new flag can be used to specify the list of different compilers for
`rdmd` to invoke when running the test suite.  This can be provided as a
single compiler name, or as a comma-separated list, e.g.

    --test-compilers=dmd,gdmd,ldmd2

The complete test suite will be run for each of the specified compilers.
If the flag is not supplied, then the test suite will be run for the
same compiler as was used to build `rdmd` (as specified by `rdmd_test`'s
`--compiler` flag).

The test for the fallback to default compiler is still invoked only with
respect to the build compiler, since by its very nature this involves a
search for only the one compiler name (determed when `rdmd` is built).