Rename compiler parameter of rdmd_test main() to buildCompiler
This is a little bit bikesheddy but helps to emphasize the distinction
between the compiler used to build rdmd, and the compiler(s) invoked by
rdmd when running the test suite.

Note that the `--compiler` flag of `rdmd_test` has been left unchanged
so as not to risk breaking changes for users.