Break out compiler fallback test into separate function of rdmd_test
If no explicit `--compiler` flag is passed, `rdmd` should first search
for its default compiler (determined by the compiler used to build it)
in the same directory that it itself lives in.

The test of this behaviour is unique inasmuch as, unlike all other tests
implemented by `rdmd_test`, it alone explicitly involves a case where
`rdmd` is searching for a single compiler name.  It must therefore be
separated out from all the other tests if we want to be able to rework
them to run with multiple different compilers (which clearly requires
use of the `--compiler` flag).

The new `runFallbackTest` function has been implemented so as to avoid
having any dependencies on global state, since this will simplify future
changes required by multi-compiler tests.