Allow setting rdmd_test test compilers when invoking posix.mak makefile
The `posix.mak` makefile has been updated to allow the user to set a new
`RDMD_TEST_COMPILERS` variable (which by default is set to the absolute
path of the compiler used to build all the D tools, including rdmd).

This can be used locally along the lines of:

    make -f posix.mak test RDMD_TEST_COMPILERS=dmd,gdmd,ldmd2

but the main point of this patch is to allow CI to start using this new
multi-compiler test functionality.  This will be added separately in a
follow-up patch.

No modifications have been made to windows makefiles, since this is not
required by Travis CI.