Set multiple rdmd test compilers in travis.sh CI script
This patch is something of a proof-of-concept for using rdmd_test's new
`--test-compilers` flag to test rdmd using multiple different compilers.

The `install.sh` script is used to install LDC (including the DMD-alike
`ldmd2` compiler interface interface), and the `make -f posix.mak test`
call is updated to specify both `dmd` and `ldmd2` as test compilers for
the `rdmd_test` test suite.  A little trickery with `find` is necessary
to work around the problem that the install location of` ldmd2` is not
added to `PATH`.

The `VERBOSE_RDMD_TEST` flag is set to ensure that the Travis build logs
will contain a full summary of the individual `rdmd` calls from the test
suite, for each of the test compilers.

This should probably not be the final form for setting up multi compiler
tests, but serves as an initial stage which should at a minimum flag up
issues like that observed with https://github.com/dlang/tools/pull/271.
While in development it has already shown up at least one `rdmd` issue
(fixed in https://github.com/dlang/tools/pull/303), so we can reasonably
hope that it will prevent others from arising in the first place.