Add standalone functions to generate code for --eval and --loop flags
The `--eval` and `--loop` flags allow the user to specify small snippets
of code that are placed inside the body of an auto-generated `main` that
is written to a file and evaluated.

The existing implementation uses two separate string concatenations for
the two different cases.  Besides the code duplication, this means that
the resulting code generation cannot be tested.

A further more subtle issue arises from the existing implementation's
insistence on adding a closing `;` to the generated code.  While this
takes care of the careless user who forgets to write one themselves, it
means that if the user does close their statements correctly, we end up
with an empty statement on the end of the code to be evaluated.  While
some D compilers don't care about this, some object:
https://github.com/dlang/tools/pull/297#issuecomment-362447334

... meaning that `rdmd` will fail to successfully evaluate the code.

This patch therefore implements two new code functions to generate the
required code.  The first, `innerEvalCode`, takes care of the innermost
code (the body of either the `main` function for `--eval`, or the loop
for `--loop`), adding a terminating `;` if one is not present, but not
doing so if one is already in place.

The second, `makeEvalCode`, takes two parameters: the `string[]` of code
snippets to be evaluated, and a boolean `Flag` indicating whether or not
this is the body of a loop (in other words, whether this code should be
generated for the `--eval` flag or the `--loop` flag).

Unittests have been provided with both to ensure that they produce the
expected outcomes.