Changed !isDir to isFile
File types are split into more than just directories and files. In this case, fullExe must be a regular file, so checking isFile removes any question about whether or not it is actually a file. The current implementation uses !isDir which could still be true even if fullExe is not a regular file, which is not the logic we want here.