posix.mak: declare all variables used by test_rdmd next to the target
This is just a small piece of cleanup to group together all the various
make variables used with the `test_rdmd` target immediately prior to the
target itself, so they are easier to find and their purpose is easier to
understand.

It should help clarify that these variables are used only with tests and
do not impact application builds.