win32.mak: ensure `clean` target removes all generated files and dirs
The `clean` target has been updated to remove the entire directory of
files generated by the build, together with all its contents.  This
matches what is done in the dmd, druntime, and phobos makefiles, and
should be much more maintainable, since it no longer relies on correct
specification of all the individual files that need removing.

As a byproduct, this fixes an oversight in the old `clean` target which
meant that `rdmd_test.exe` was not being removed.

The `ROOT_OF_THEM_ALL` variable has been renamed to `GENERATED` to avoid
any nasty consequences for downstream users who might be overriding the
former in their build scripts.