rdmd_test: simplify flag for specifying default compiler
Since it seems there are some use-cases where it is unavoidable to run
`rdmd_test` directly, it's nice if we can reduce the amount of typing
required to do so.

Besides the rdmd executable itself, the only other obligatory argument
to the test suite is the default compiler expected to be used by rdmd.
This patch shortens the long option to `--default-compiler` and allows
an equivalent single-character `-D` option, so that usage can now be of
the form:

    rdmd_test -D <default-compiler> <path-to-rdmd-executable>

The posix.mak and win32.mak makefiles have been updated accordingly.