rdmd: split hardcoded config variables into separate rdmd.config module
This separates out all the hardcoded settings, some derived from either
platform or compiler, and groups them inside a new `RDMDConfig` struct
which can serve effectively as a namespace.

This should ensure that these global variables are easy to identify
wherever they occur in the codebase.  Note that it may be possible to
make some of them non-global as the codebase is further refactored.

Documentation has been added for the struct and all its fields.