rdmd: split arguments-derived globals into separate rdmd.args module
This separates out all the global variables derived one way or another
from the application's command-line arguments, and groups them inside a
new `RDMDGlobalArgs` struct which can serve effectively as a namespace.

This is obviously rather horribly intrusive, but wait!  There's a method
to this madness ;-)

In the short term, this should ensure that all uses of these globals are
easy to find and identify.  In the long run, there are three major goals
for the new module:

  * encapsulate all arguments-parsing code in a single module

  * ensure that all directly arguments-derived values are grouped
    into a single data structure, whose values are written only
    from within this module

  * draw clear lines between values derived from arguments, versus
    values that result from combinations of arguments

In the case of the last, consider e.g. the way that the `exe` value may
be overwritten (or set) depending on various combinations of flags, or
the way in which it can be dual-purposed to first contain the directory
path specified by the `-od` flag, and then later in the program the path
to the executable to be built.

This is a nastily complex way of doing things (easily revealed thanks to
the `RDMDGlobalArgs` namespacing): it would be much easier to understand
and maintain the code if instead of multipurposing one global, the path
to the build executable was a separate variable whose value was set from
a (pure) function taking the parsed command-line arguments as input.

More generally, we can expect that rdmd's code will become much simpler
to reason about if we have a clean separation between parsing command
line arguments, versus decisions that get made based on what the parsed
values are.

This patch ensures that the existing globals cannot readily be ignored.
Follow-up patches will put in place an alternative system for parsing
and using the command-line arguments, which should ensure the same end
results but with much cleaner separation between different decisions.