rdmd_test: Replace use of assert with enforce
Avoid any potential false positives resulting from rdmd_test being
accidentally built with -release.

Avoiding this situation is paramount, with the alternative being a
clunky runtime test to detect that rdmd_test was built with -release,
so this solution is preferable, even considering the semantic loss of
the distinction between assert/enforce and their existing usage.